requirements.txt
numpy
pandas
flask
psutil
redis
asyncio
pymongo
motor
tensorflow
scikit-learn
python-dotenv
